# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Notebooks",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/v1/notebooks/photo": {
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Upload photo",
                "operationId": "5b27f7296a7f655ae2ca4b01d2ef3c45",
                "responses": {
                    "201": {
                        "description": "Photo uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred while uploading the photo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebooks": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get all notebooks",
                "operationId": "16477c32e83a576344232ede3d1e415d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "father_name": {
                                                "type": "string"
                                            },
                                            "company_name": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "phone": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "birth_date": {
                                                "type": "string",
                                                "format": "date",
                                                "nullable": true
                                            },
                                            "image_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Create a notebook",
                "operationId": "82659f0e438b64031e8076436bf9cca3",
                "requestBody": {
                    "description": "Notebook data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "father_name": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "ABC Inc.",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1234567890",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01",
                                        "nullable": true
                                    },
                                    "image_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notebook created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "last_name": {
                                            "type": "string"
                                        },
                                        "first_name": {
                                            "type": "string"
                                        },
                                        "father_name": {
                                            "type": "string"
                                        },
                                        "company_name": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "birth_date": {
                                            "type": "string",
                                            "format": "date",
                                            "nullable": true
                                        },
                                        "image_id": {
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "An error occurred while creating the notebook",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebooks/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get a notebook by id",
                "operationId": "0cc9df8c674c0c23798c31b8f705cc00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "last_name": {
                                            "type": "string"
                                        },
                                        "first_name": {
                                            "type": "string"
                                        },
                                        "father_name": {
                                            "type": "string"
                                        },
                                        "company_name": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "birth_date": {
                                            "type": "string",
                                            "format": "date",
                                            "nullable": true
                                        },
                                        "image_id": {
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notebook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Delete a notebook",
                "description": "Delete a notebook by ID",
                "operationId": "684e2dd08b9b70b80641d4c50f0246a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the notebook to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Notebook deleted successfully"
                    },
                    "404": {
                        "description": "Notebook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Update a notebook",
                "operationId": "6f4c4de54da5f202ac1184d81bea0262",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data to update the notebook",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "father_name": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notebook updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "last_name": {
                                            "type": "string"
                                        },
                                        "first_name": {
                                            "type": "string"
                                        },
                                        "father_name": {
                                            "type": "string"
                                        },
                                        "company_name": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "birth_date": {
                                            "type": "string",
                                            "format": "date",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notebook not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}